+++ src/SomeDL/core/processor.py
@@ -1,6 +1,7 @@
 import json
 import time
 import re
+import random
 import traceback
 
 import threading
@@ -36,7 +37,7 @@ def process_song_list_concurrent(song_list_queue: queue.Queue, oneshot: bool = T
     if oneshot:
         song_list_queue.put(_DONE)
 
-    # === Downloader thread ===
+    # === Metadata fetching thread ===
     def thread_fetch_metadata(song_list_queue):
         metadata_list = []
         
@@ -137,97 +138,6 @@ def process_song_list_concurrent(song_list_queue: queue.Queue, oneshot: bool = T
             download_queue.put(_DONE)
 
 
-    # === Metadata fetching thread ===
-    def thread_fetch_metadata_old(songs_list):
-        metadata_list = []
-        
-        index = 0
-        length = len(songs_list)
-
-        for item in songs_list:
-            index += 1
-
-            if item.get("text_query"):
-                label = f'{index}/{length} {item.get("text_query")}'
-            else:
-                label = f'{index}/{length} {item.get("artist_name")} - {item.get("song_title")}'
-            item["label"] = label
-            
-            try:
-                
-                # === clean metadata ===
-                clean_success = metadata_type_cleaner(item)
-
-                if not clean_success:
-                    console.warning(f'Song could not be downloaded, error on initial lookup.', item["label"])
-                    console.finish(item.get("label"), console.Download_status.FAILED)
-                    with console.thread_lock:
-                        failed_list.append(item)
-                    continue
-
-                # === Create the real label after cleaning metadata. ===
-                label = f'{index}/{length} {item.get("artist_name")} - {item.get("song_title")}'
-                item["label"] = label
-
-                # === Check if the song was already in the queue ===
-                # --- (Necessary for when the same song is input twice, but is not yet downloaded)
-                if any(d.get("artist_name") == item.get("artist_name") and d.get("song_title") == item.get("song_title") for d in metadata_list):
-                    console.info("This input is a duplicate", item["label"])
-                    console.finish(item.get("label"), console.Download_status.ALREADY_DOWNLOADED)
-
-                    with console.thread_lock:
-                        already_downloaded_list.append(item)
-                    continue
-
-
-                # === fetch metadata ===
-                
-                metadata = fetch_metadata(item, metadata_list)
-
-                if not metadata:
-                    console.warning("Song could not be downloaded", item["label"])
-                    console.finish(item.get("label"), console.Download_status.FAILED)
-                    with console.thread_lock:
-                        failed_list.append(item)
-                    continue
-
-                if metadata == "already_downloaded":
-                    console.finish(item.get("label"), console.Download_status.ALREADY_DOWNLOADED)
-                    
-                    with console.thread_lock:
-                        already_downloaded_list.append(item)
-                    continue
-
-                metadata_list.append(metadata) # --- Used for known_metadata
-
-                if config["download"]["disable_download"]:
-                    console.update(label, "disable_download", console.Status.SKIPPED)
-                    console.notice("Not downloading song because disable-download is set", item["label"])
-                    console.finish(label, console.Download_status.DOWNLOAD_DISABLED)
-
-                    with console.thread_lock:
-                        metadata_success_list.append(metadata)
-                    continue
-
-                console.update(label, "wait_queue", console.Status.ACTIVE, "Queuing for download")
-
-                download_queue.put(metadata) # ---- blocks when queue is full
-
-                console.update(label, "wait_queue", console.Status.ACTIVE, "Queuing for download (in queue)")
-                
-            except Exception as e:
-                console.error("A critical exception occurred when fetching the metadata for this song! If retrying does not help, please notify the program maintainer at https://github.com/ChemistryGull/SomeDL. Error:", label)
-                console.error(e, label)
-                console.finish(label, console.Download_status.FAILED)
-                with console.thread_lock:
-                    failed_list.append(item)
-
-        for _ in range(NUM_DOWNLOADERS):
-            download_queue.put(_DONE)
-            
-
-
-
     # === Downloader thread ===
     def thread_downloader(thread_id: int) -> None:
         while True:
@@ -238,7 +148,14 @@ def process_song_list_concurrent(song_list_queue: queue.Queue, oneshot: bool = T
             label = None
             try:
                 timer_start = time.time()
-                label = metadata.get("label")    
+                label = metadata.get("label")
+
+                sleep_time = config["download"]["sleep"] + random.uniform(0, 5)
+
+                if config["download"]["sleep"]:
+                    console.update(label, "wait_queue", console.Status.ACTIVE, f'yt-dlp: Sleeping for {round(sleep_time, 1)} seconds.')
+                    time.sleep(sleep_time) # --- Adds random amount of seconds 0-5 to the sleep timer
+                
                 console.update(label, "wait_queue", console.Status.HIDE)
 
                 console.update(label, "downloading", console.Status.ACTIVE, "yt-dlp: Preparing download")